+++ src/spring-cloud/azext_spring_cloud/_app_validator.py
@@ -7,11 +7,13 @@
 
 from azure.cli.core.azclierror import InvalidArgumentValueError
 from msrestazure.azure_exceptions import CloudError
+from azure.core.exceptions import (ResourceNotFoundError)
 from ._client_factory import cf_spring_cloud
 
 
 # pylint: disable=line-too-long,raise-missing-from
 NO_PRODUCTION_DEPLOYMENT_ERROR = "No production deployment found, use --deployment to specify deployment or create deployment with: az spring-cloud app deployment create"
+NO_PRODUCTION_DEPLOYMENT_SET_ERROR = "This app has no production deployment, use \"az spring-cloud app deployment create\" to create a deployment and \"az spring-cloud app set-deployment\" to set production deployment."
 
 
 def fulfill_deployment_param(cmd, namespace):
@@ -22,6 +24,24 @@ def fulfill_deployment_param(cmd, namespace):
         namespace.deployment = _ensure_deployment_exist(client, namespace.resource_group, namespace.service, namespace.name, namespace.deployment)
     else:
         namespace.deployment = _ensure_active_deployment_exist_and_get(client, namespace.resource_group, namespace.service, namespace.name)
+
+
+def active_deployment_exist(cmd, namespace):
+    if not namespace.name or not namespace.service or not namespace.resource_group:
+        return
+    client = cf_spring_cloud(cmd.cli_ctx)
+    deployment = _get_active_deployment(client, namespace.resource_group, namespace.service, namespace.name)
+    if not deployment:
+        raise InvalidArgumentValueError(NO_PRODUCTION_DEPLOYMENT_SET_ERROR)
+
+
+def active_deployment_exist_under_app(cmd, namespace):
+    if not namespace.app or not namespace.service or not namespace.resource_group:
+        return
+    client = cf_spring_cloud(cmd.cli_ctx)
+    deployment = _get_active_deployment(client, namespace.resource_group, namespace.service, namespace.app)
+    if not deployment:
+        raise InvalidArgumentValueError(NO_PRODUCTION_DEPLOYMENT_SET_ERROR)
 
 
 def _ensure_deployment_exist(client, resource_group, service, app, deployment):
@@ -39,5 +59,8 @@ def _ensure_active_deployment_exist_and_get(client, resource_group, service, nam
 
 
 def _get_active_deployment(client, resource_group, service, name):
-    deployments = client.deployments.list(resource_group, service, name)
-    return next(iter(x for x in deployments if x.properties.active), None)
+    try:
+        deployments = client.deployments.list(resource_group, service, name)
+        return next(iter(x for x in deployments if x.properties.active), None)
+    except ResourceNotFoundError:
+        raise InvalidArgumentValueError('Deployments not found under App {}'.format(name))