+++ src/spring-cloud/azext_spring_cloud/custom.py
@@ -462,12 +462,6 @@ def _default_deployment_resource_builder(cpu, memory, env, jvm_options, runtime_
     return deployment_resource
 
 
-def _check_active_deployment_exist(client, resource_group, service, app):
-    active_deployment_name = client.apps.get(resource_group, service, app).properties.active_deployment_name
-    if not active_deployment_name:
-        logger.warning(NO_PRODUCTION_DEPLOYMENT_SET_ERROR)
-
-
 def app_update(cmd, client, resource_group, service, name,
                assign_endpoint=None,
                deployment=None,
@@ -633,12 +627,9 @@ def app_get(cmd, client,
             service,
             name):
     app = client.apps.get(resource_group, service, name)
-    deployment_name = app.properties.active_deployment_name
-    if deployment_name:
-        deployment = client.deployments.get(
-            resource_group, service, name, deployment_name)
-        app.properties.active_deployment = deployment
-    else:
+    deployments = client.deployments.list(resource_group, service, name)
+    app.properties.active_deployment = next((x for x in deployments if x.properties.active), None)
+    if not app.properties.active_deployment:
         logger.warning(NO_PRODUCTION_DEPLOYMENT_SET_ERROR)
 
     return app
@@ -762,7 +753,6 @@ def app_tail_log(cmd, client, resource_group, service, name,
 
 
 def app_identity_assign(cmd, client, resource_group, service, name, role=None, scope=None):
-    _check_active_deployment_exist(client, resource_group, service, name)
     app_resource = models_20210601preview.AppResource()
     identity = models_20210601preview.ManagedIdentityProperties(type="systemassigned")
     properties = models_20210601preview.AppResourceProperties()
@@ -820,7 +810,6 @@ def app_identity_remove(cmd, client, resource_group, service, name):
 
 
 def app_identity_show(cmd, client, resource_group, service, name):
-    _check_active_deployment_exist(client, resource_group, service, name)
     app = client.apps.get(resource_group, service, name)
     return app.identity
 
@@ -849,11 +838,6 @@ def app_set_deployment(cmd, client, resource_group, service, name, deployment):
 
 
 def app_unset_deployment(cmd, client, resource_group, service, name):
-    active_deployment_name = client.apps.get(
-        resource_group, service, name).properties.active_deployment_name
-    if not active_deployment_name:
-        raise CLIError(NO_PRODUCTION_DEPLOYMENT_SET_ERROR)
-
     # It's designed to use empty string for active_deployment_name to unset active deployment
     properties = models_20210601preview.AppResourceProperties(active_deployment_name="")
 
@@ -1279,12 +1263,10 @@ def config_repo_list(cmd, client, resource_group, name):
 
 
 def binding_list(cmd, client, resource_group, service, app):
-    _check_active_deployment_exist(client, resource_group, service, app)
     return client.bindings.list(resource_group, service, app)
 
 
 def binding_get(cmd, client, resource_group, service, app, name):
-    _check_active_deployment_exist(client, resource_group, service, app)
     return client.bindings.get(resource_group, service, app, name)
 
 
@@ -1298,7 +1280,6 @@ def binding_cosmos_add(cmd, client, resource_group, service, app, name,
                        database_name=None,
                        key_space=None,
                        collection_name=None):
-    _check_active_deployment_exist(client, resource_group, service, app)
     resource_id_dict = parse_resource_id(resource_id)
     resource_type = resource_id_dict['resource_type']
     resource_name = resource_id_dict['resource_name']
@@ -1332,7 +1313,6 @@ def binding_cosmos_update(cmd, client, resource_group, service, app, name,
                           database_name=None,
                           key_space=None,
                           collection_name=None):
-    _check_active_deployment_exist(client, resource_group, service, app)
     binding = client.bindings.get(resource_group, service, app, name).properties
     resource_id = binding.resource_id
     resource_name = binding.resource_name
@@ -1360,7 +1340,6 @@ def binding_mysql_add(cmd, client, resource_group, service, app, name,
                       key,
                       username,
                       database_name):
-    _check_active_deployment_exist(client, resource_group, service, app)
     resource_id_dict = parse_resource_id(resource_id)
     resource_type = resource_id_dict['resource_type']
     resource_name = resource_id_dict['resource_name']
@@ -1383,7 +1362,6 @@ def binding_mysql_update(cmd, client, resource_group, service, app, name,
                          key=None,
                          username=None,
                          database_name=None):
-    _check_active_deployment_exist(client, resource_group, service, app)
     binding_parameters = {}
     binding_parameters['username'] = username
     binding_parameters['databaseName'] = database_name
@@ -1399,7 +1377,6 @@ def binding_mysql_update(cmd, client, resource_group, service, app, name,
 def binding_redis_add(cmd, client, resource_group, service, app, name,
                       resource_id,
                       disable_ssl=None):
-    _check_active_deployment_exist(client, resource_group, service, app)
     use_ssl = not disable_ssl
     resource_id_dict = parse_resource_id(resource_id)
     resource_type = resource_id_dict['resource_type']
@@ -1426,7 +1403,6 @@ def binding_redis_add(cmd, client, resource_group, service, app, name,
 
 def binding_redis_update(cmd, client, resource_group, service, app, name,
                          disable_ssl=None):
-    _check_active_deployment_exist(client, resource_group, service, app)
     binding = client.bindings.get(resource_group, service, app, name).properties
     resource_id = binding.resource_id
     resource_name = binding.resource_name
@@ -1866,7 +1842,6 @@ def domain_bind(cmd, client, resource_group, service, app,
                 domain_name,
                 certificate=None,
                 enable_end_to_end_tls=None):
-    _check_active_deployment_exist(client, resource_group, service, app)
     properties = models.CustomDomainProperties()
     if certificate is not None:
         certificate_response = client.certificates.get(resource_group, service, certificate)
@@ -1899,12 +1874,10 @@ def _update_app_e2e_tls(cmd, resource_group, service, app, enable_end_to_end_tls
 
 
 def domain_show(cmd, client, resource_group, service, app, domain_name):
-    _check_active_deployment_exist(client, resource_group, service, app)
     return client.custom_domains.get(resource_group, service, app, domain_name)
 
 
 def domain_list(cmd, client, resource_group, service, app):
-    _check_active_deployment_exist(client, resource_group, service, app)
     return client.custom_domains.list(resource_group, service, app)
 
 
@@ -1912,7 +1885,6 @@ def domain_update(cmd, client, resource_group, service, app,
                   domain_name,
                   certificate=None,
                   enable_end_to_end_tls=None):
-    _check_active_deployment_exist(client, resource_group, service, app)
     properties = models.CustomDomainProperties()
     if certificate is not None:
         certificate_response = client.certificates.get(resource_group, service, certificate)