+++ src/spring-cloud/azext_spring_cloud/tests/latest/test_asc_app_validator.py
@@ -7,7 +7,9 @@ import copy
 from argparse import Namespace
 from azure.cli.core.azclierror import InvalidArgumentValueError
 from msrestazure.azure_exceptions import CloudError
-from ..._app_validator import (fulfill_deployment_param)
+from azure.core.exceptions import ResourceNotFoundError
+from ..._app_validator import (fulfill_deployment_param, active_deployment_exist, active_deployment_exist_under_app)
+
 
 try:
     import unittest.mock as mock
@@ -35,6 +37,58 @@ def _get_deployment(resource_group, service, app, deployment, active):
     resource.properties = mock.MagicMock()
     resource.properties.active = active
     return resource
+
+
+class TestActiveDeploymentExist(unittest.TestCase):
+    @mock.patch('azext_spring_cloud._app_validator.cf_spring_cloud', autospec=True)
+    def test_deployment_found(self, client_factory_mock):
+        client = mock.MagicMock()
+        client.deployments.list.return_value = [
+            _get_deployment('rg1', 'asc1', 'app1', 'green1', False),
+            _get_deployment('rg1', 'asc1', 'app1', 'default', True),
+        ]
+        client_factory_mock.return_value = client
+
+        ns = Namespace(name='app1', service='asc1', resource_group='rg1', deployment=None)
+        active_deployment_exist(_get_test_cmd(), ns)
+
+    @mock.patch('azext_spring_cloud._app_validator.cf_spring_cloud', autospec=True)
+    def test_deployment_without_active_exist(self, client_factory_mock):
+        client = mock.MagicMock()
+        client.deployments.list.return_value = [
+            _get_deployment('rg1', 'asc1', 'app1', 'green1', False)
+        ]
+        client_factory_mock.return_value = client
+
+        ns = Namespace(name='app1', service='asc1', resource_group='rg1', deployment=None)
+        with self.assertRaises(InvalidArgumentValueError) as context:
+            active_deployment_exist(_get_test_cmd(), ns)
+        self.assertEqual('This app has no production deployment, use \"az spring-cloud app deployment create\" to create a deployment and \"az spring-cloud app set-deployment\" to set production deployment.', str(context.exception))
+
+    @mock.patch('azext_spring_cloud._app_validator.cf_spring_cloud', autospec=True)
+    def test_no_deployments(self, client_factory_mock):
+        client = mock.MagicMock()
+        client.deployments.list.return_value = []
+        client_factory_mock.return_value = client
+
+        ns = Namespace(name='app1', service='asc1', resource_group='rg1', deployment=None)
+        with self.assertRaises(InvalidArgumentValueError) as context:
+            active_deployment_exist(_get_test_cmd(), ns)
+        self.assertEqual('This app has no production deployment, use \"az spring-cloud app deployment create\" to create a deployment and \"az spring-cloud app set-deployment\" to set production deployment.', str(context.exception))
+
+    @mock.patch('azext_spring_cloud._app_validator.cf_spring_cloud', autospec=True)
+    def test_app_not_found(self, client_factory_mock):
+        client = mock.MagicMock()
+        resp = mock.MagicMock()
+        resp.status_code = 404
+        resp.text = '{"Message": "Not Found"}'
+        client.deployments.list.side_effect = ResourceNotFoundError(resp, error='App not found.')
+        client_factory_mock.return_value = client
+
+        ns = Namespace(name='app1', service='asc1', resource_group='rg1', deployment=None)
+        with self.assertRaises(InvalidArgumentValueError) as context:
+            active_deployment_exist(_get_test_cmd(), ns)
+        self.assertEqual('Deployments not found under App app1', str(context.exception))
 
 
 class TestFulfillDeploymentParameter(unittest.TestCase):