+++ src/aks-preview/HISTORY.md
@@ -2,6 +2,11 @@
 
 Release History
 ===============
+0.5.47
+++++++
+* Add support for IPv4/IPv6 dual-stack networking AKS clusters
+* `az aks create --pod-cidrs --service-cidrs --ip-families --load-balancer-managed-outbound-ipv6-count`
+
 0.5.46
 ++++++
 * Update to use 2021-10-01 api-version